Refactor for unit tests of queue
