+++ src/ipc/ipc.hpp
@@ -28,4 +28,49 @@ public:
     // Generate an instance IPC based on the desired method, role and parameters
     static std::unique_ptr<IPC> get_ipc(Method method, Role role, OptArgs& args);
 };
+
+class MQRead: public IPC{
+    private:
+        // Used to store message already read
+        std::vector<byte> buffer;
+        // If true the next read should read from buffer
+        bool buffer_busy;
+        // Becomes true when the writer sends an high priority message
+        // stating that it finished sending. If it's true an empty queue
+        // means the process received the whole file
+        bool writer_finished;
+        std::string queue_name;
+        std::string lock_name;
+    public:
+        MQRead(Method Method, Role role, OptArgs& args);
+        ~MQRead();
+
+        size_t buff_size() final;
+
+        /* Call after checking with ready(), will return a vector of bytes
+         * Max read is equal to the return value of buff_size
+         */
+        std::vector<byte> receive(size_t max_read) final;
+
+        // Not implemented for receiver
+        int send(std::vector<byte> payload) final;
+
+        /* Returns true if messages are still expected, if false the reader 
+         * process can shut down
+         */
+        bool ready() final;
+};
+
+class MQWrite: public IPC{
+    public:
+        MQWrite(Method Method, Role role, OptArgs& args);
+
+        size_t buff_size() final;
+
+        std::vector<byte> receive(size_t max_read) final;
+
+        int send(std::vector<byte> payload) final;
+
+        bool ready() final;
+};
 #endif
