+++ tests/BUILD
@@ -4,6 +4,6 @@ cc_test(
   srcs = ["hello_test.cc"],
   deps = [
     "@com_google_googletest//:gtest_main",
-    "//src:helplib"
+    "//src:lib_ocp"
     ],
 )
