+++ tests/hello_test.cc
@@ -1,9 +1,11 @@
 #include <gtest/gtest.h>
 #include "src/help.hpp"
+#include "src/ipc/ipc.hpp"
 
 // Demonstrate some basic assertions.
 TEST(HelloTest, BasicAssertions) {
   // Expect two strings not to be equal.
+
   EXPECT_STRNE("hello", "world");
   // Expect equality.
   EXPECT_EQ(mult(), RES);
